requirements.txt
# TechFlow AI Recruiter - Dependencies

# LangChain Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# OpenAI
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# MongoDB
pymongo>=4.6.0

# Web Interface
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment
python-dotenv>=1.0.0

# PDF Processing
pypdf>=3.0.0
pdfplumber>=0.10.0

# Tokenization
tiktoken>=0.5.0

# Evaluation & Metrics
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
